Update level to match changes to plugin (SHA 3fac075)
The main car demo level was originally designed to compensate for the different coordinate system and scaling of Unity.
Changes made to the plugin on May 17 (SHA 3fac075) to also compensate for the different coordinate system caused the rotation of objects to be mirrored.
This update adjusts the scene in response to the plugin changes.

- Update to build 697 of the plugin