Decouple UI widget from Gamemode
Previously, the UI widget was defined as HUD class of the Gamemode. However, since this uses the Cavrnus Gamemode, that would break every time the plugin is updated.

The UI is now spawned by the CSC actor in the level. Both options are still available, but this one works better for continuous testing and demonstrations.